+++ src/main/java/roomescape/controller/ReservationController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import roomescape.domain.Reservation;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.SaveReservationRequest;
+import roomescape.service.dto.ReservationResponse;
+import roomescape.service.dto.SaveReservationRequest;
 import roomescape.service.ReservationCreateService;
 import roomescape.service.ReservationDeleteService;
 import roomescape.service.ReservationFindService;
